host: make Host.run() cancellable
The real purpose of this is to replace multiprocessing.Process()
with

    thread = common.FutureThread(
        lambda th: rsh.run(
            "echo foo; sleep 13",
            cancellable=th.cancellable,
        ),
        start=True,
    )
    ...
    res = thread.join_and_result(cancel=True)

With multiprocessing.Process() we can of course always kill the external
process. But IMO multiprocessing.Process() is problematic, because it
either uses fork() or spans another Python instance. Both can have
surprising effects. Instead, make Host.run() suitable to run via a
Thread (which requires cancellation).

For LocalHost.run() we send a SIGTERM and expect the process to quit. It
is not supported that the process refuses to terminate. There is only a
timeout to log an error and send a SIGKILL. But that is only to catch
problems, it's not how you use this. A process that refuses to die will
leak as a zombie and we cannot collect the exit code. Only spawn
processes that you know what they are doing, and only cancel them if
they let themselves be terminated with SIGTERM (like most processes
should).

For RemoteHost.run(), cancellation is cumbersome. SSH doesn't really
have a way to kill the process. Instead, in that case we run a shell
script which watches stdin. When stdin closes, the script kills the
command.