host: let host._run() log in the context of the call
Host.run() has additional context about the logging. The Host._run()
implementation should log in the same way.

There are different ways how to achieve that. For example, we could pass
down the necessary information for how to log the message (that includes
at least the log_prefix and log_id). We could bundle that information in
a struct, so we only pass one piece of data around. Instead, there is a
now an internal handle_log callback argument. That callback is defined
at one place only, and you can read there what it does. The callers of the
callback also don't need to know anything else, except that they want to
log a message and the desired level.