host: disable log_level when log_level_result is set
In most cases, callers don't set the log level. If they set the
log_level argument, then that is used both for the message when the
command starts (the preamble message) and when it completes.

However, if the caller only sets the log_level_result argument, we can
assume that by default they don't want to see the preamble message. In
that case, automatically disable log_level.

Also, if the premable message is not shown but only the log_level_result
message, then don't show "└──>". Instead, just show ">".

This mode of logging (where we only log the result in a one liner) makes
the most sense for short-running commands, where it's not very
intersting to see when the command starts.