host: disable log_level when log_level_result is set
In most cases, callers don't set the log level. If they set it, that's
it.

However, if the caller only sets the log_level_result, we can assume
that by default they don't want to see the preamble message. In that
case, disable log_level.

Also, if the premable message is not shown but only the log_level_result
message, then don't show "└──>". Instead, just show ">".