host: use common.KW_ONLY_DATACLASS in host.py
So far this was avoided, because it wasn't clear whether "host" module
should depend on "common" module or vice versa.

That was also, because in the past netdev.ip_addrs() was inside the
"common" module, so back then, "host" could not have imported "common"
(but the other way around).

Instead, clarify that "host" very much can use "common" module (and
never the other way around).

At this point, use common.KW_ONLY_DATACLASS instead of reimplementing
it.