common: add StructParseBase._owner_reference field and DeferredReference class
This allows the user to hook up an owner/parent reference.

The StructParseBase objects represent parts of a YAML file. It can be
useful to go from the deeper nested node to the parent. However,
these objects are also all immutable. So we cannot create the parent
instance first, then the object (setting the parent/owner), and finally
let the parent track the object.

Instead, we first create the object (that has a _owner_reference
initialized to None). Then the parent/owner will be created and track
the object. Finally the parent sets itself as the owner_reference and
closes the loop.  Note that almost all involved objects are immutable
once created, except DeferredReference.init(). And DeferredReference
only allows to call get() after the value was initialized. Wo while it's
not initialized from the beginning, every caller can only ever see the
same value (or get an exception).

Also use a weakref. The point is to avoid cyclic references which would
hamper destruction of the owner and object referencing each other.

Also, use a (ClassVar) lock so this is thread-safe. This lock is per
class to avoid an locking instance for each object. It gives
thread-safety for the unusual case, but with a smaller allocation
overhead than having one lock per object.