common: accept pathlib.Path argument in common.path_norm()
Also accept pathlib.Path arguments and return the result in the same
kind.

Also accept None, if given, it returns None. It's convenient to the
caller to accept Optional[str] and propagate the missing value. If the
provided argument is already known to be not of type None, then mypy
will still understand that this never returns None. So this is only
useful for callers that pass in an Optional[str]. They safely get an
Optional[str] back.