common: add StructParseParseContext for packaging parse arguments
With the "structparse" helpers, we are expected to have parse() methods,
that take commonly 3 parameters ["yamlidx", "yamlpath", "arg"].

Pack them together in a StructParseParseContext object. It avoids some
typing. It also makes it clearer to the user how to use the pattern when
implementing the parse methods.

Also, structparse_pop_objlist() etc. have a callback argument to
construct the data. That callback is also a parse() method, and will
also receive the same data (StructParseParseContext). It's convenient to
bundle them together.