common: fix @iter_{listify,tuplify,dictify} for Python 3.9
typing.ParamSpec only exists in Python 3.10. For 3.9 support we need a
less powerful alternative that does not correctly detect the types of
the decorated function.

This is acceptable, because it works at runtime with Python 3.9. And
while having less strict type checks with 3.9, once you check with 3.10,
you get the strict checks.

Fixes: d114e7120d65 ('common: add @iter_{listify,tuplify,dictify} decorators')