common: add common.iter_eval_now() helper
We want that our function have Iterable[T] parameters, so that they can
be used with various kinds of arguments, including generator
expressions.

However, such iterables can only be iterated once. Inside the
implementation we want to make sure that we evaluate the entire iterable
exactly once.

Add common.iter_eval_now(). This returns lists/tuples as-is, however for
other iterables it will create a tuple() to evaluate and cache the
iterable for repeated use.