kyaml: add ktoolbox.kyaml module with dump helpers
Pyyaml's yaml.dump() does not by default generate the format we want.
For one, we need default_flow_style=False. But also, it generates YAML
like:

  foo:
  - id: 1

when we would want:

  foo:
    - id: 1

We can get that by setting the Dumper argument. However, doing that each
time is cumbersome, so add kyaml.dump() and kyaml.dumps() helpers to do
it the way we want.

This is not part of ktoolbox.common, because PyYaml should be an
optional dependency.