common: add json_dump() helper
The json package with its dump()/dumps() methods is useful and flexible.
By default it does not pretty print the output. That makes sense.

However, sometimes we use dataclasses that we want to persist to file.
Note that those detaclasses won't be user configuration, as for user
configuration we would instead write YAML. However, for generated data
it makes sense to convert to dict (e.g. dataclasses.asdict()) and then
store as JSON. We can then parse such files with
common.dataclass_from_{dict,json,file}().

When we write such files, we want to wrap lines (indent=2). We also want
a trailing newline, so it looks nicer in the editor.

Add a utlity that performs this in exactly the way we want it. While
its implementation is simple, it defines the format that we are going to
write for such cases.