common: add use_or_open() helper context manager
It seems useful that functions that can write a file, alternatively
accepts a file object as parameter.

In those cases, the implementation of such a function will check the
type of the parameter, and either call open() or use the provided file
object.

Making sure to close the file if we opened is is a bit cumbersome. Add a
new helper context manager use_or_open() for that.