common: add "check_ctx" parameter to structparse_pop_{str,int,float,bool,obj}()
As it was previously, the caller could have called

    val = common.structparse_pop_str(
        varg.for_key("my_key"),
        check=lambda val: _check(val)
    )

The problem was, inside the callback we didn't know the key and other
context. Instead the caller could have specify the parameter twice:

    val = common.structparse_pop_str(
        varg.for_key("my_key"),
        check=lambda val: _check(val, "my_key")
    )

Add a new "check_ctx" callback, which gets the StructParsePopContext.
That is useful for the callee to know what key is parsed and for example
to raise an exception via pargs.value_error().