black: don't depend on fixed black==24.4.2 version
With current `pip install black` we get black==25.1.0. This is what we
get in the github actions too (if we no longer use a fixed version).

Note that on Fedora 41, we get 24.4.2. This is relevant because I use
that, and it is cumbersome if my black version disagrees with the check
in the github action.

Turns out, after reformatting the code with black==25.1.0, the code is
now compatible with 25.1.0 and 24.4.2.

In this case, it's undesirable that we tie to a certain version of
black. Drop the version restriction.