common: accept Serial.expect(pattern=None) for polling data
We have Serial.read_all() which can read all data from the serial
port. However, that always returns if there is nothing more to
read.

We have Serial.expect(), which can read_all() in a loop, and wait for an
expected pattern of a timeout (on timeout, an exception will be raised).

Reading all for a certain time is useful, even if we don't expect a
pattern. For example, Serial.read_all() (and indirectly Serial.expect())
reads all data and logs(!) it. If you want to read and log all data from
the port for a certain time, you can now use Serial.expect(None, timeout=5).

While at it, in Serial.send() don't call time.sleep(). Instead call
Serial.expect(pattern=None). This polls the serial port and we get log
message about what is received, when it is received. Note that whatever
we read, is appended to the internal _buffer, so it's still available
for the next expect().