common: fix Serial to correctly handle UTF-8 characters
Previously, Serial would read binary data, convert it to string and
append it to an internal string buffer.

However, as we read chunks of data, UTF-8 characters might be split
in the middle. In that case, the concatenated string would contain
two surrogate escaped characters, and not the proper UTF-8 character.

Fix that. Now, track the internal buffer as binary.