host: add host.Host.run_in_thread()
We already have common.FutureThread helper class for various purposes
when we want to run a thread and collect the result of a callback.

A common usage is to run host.Host.run() on another thread.

For that, add a convenience function host.Host.run_in_thread() which
combines those two and returns a newly created FutureThread.