host: create new SSHClient for each login retry
For a login, SSHClient spawns a new thread. It seems, if a previous
login fails, that old thread is not terminated right away. By retrying
multiple logins with the same client instance, the threads (and maybe
other resources) pile up.

Instead, create a new client instance for each retry. This avoids
the problem.