common: add FutureThread.cancel()
This is similar to FutureThread.cancellable.cancel().

The difference is that if the FutureThread did not yet create a
cancellable instance (which is created lazy on first use), then
FutureThread.cancel() will not create it. Instead it just remembers that
we are being cancelled until somebody requests the cancellable instance.