commit: add FutureThread.get_cancellable() method
FutureThread.cancellable is a property that creates the instance lazily
on first call. From point of the user, it looks as if the cancellable
property is immutable (they never get to see another value or None).

Creating a cancellable has an overhead. If an upper layer function
passes down a cancellable, so that lower layers may get it, they will
already need to create the cancellable instance.

However, it may be not necessary that the lower layer ever needs the
cancellable, so it is created unnecessarily.

Add FutureThread.get_cancellable() which is a callable, that can be
passed down instead. This way, the caller decides when to use and create
the cancellable.