host: workaround potential race in Host.run_in_thead() with start+add_to_thread_list
Granted, it is unlikely to hit. But don't have races.

We should start() the thread at the same time when adding it to the
thread list.

Previously, it was theoretically possible that the thread was already
running at a time, when the thread was not added to the thread list yet.