common: add run_main() helper
This replaces the common pattern:

    if __name__ == "__main__":
        try:
            main()
        except Exception:
            traceback.print_exc()
            sys.exit(70)

with:

    if __name__ == "__main__":
        ktoolbox.common.run_main(main)

Error code 70 is chosen as default exit code because that is EX_SOFTWARE
from <sysexits.h>. In any case, proper tools should handle real
application errors and not let Exception escape from their main()
function. If they do, it is either a bug or a case that they
intentionally don't handle (and we treat it like a bug). Returning 70
seems sensible in that case, as you can then clearly recognize it.