common: better handle invalid loglevel in KTOOLBOX_LOGFILE
An invalid value in the environment variables should not crash the program.
So we need to do something about invalid logging levels. Previously, we would
parse

  "bogus:xx" -> ("xx", None, True)

It seems bad to ignore the invalid part. More likely, the user intended this
to be part of the filename. So now we get

  "bogus:xx" -> ("bogus:xx", None, True)