common: use logger for common.run_main()
Accept a logger argument. If given, use that to log the traceback. The
benefit is that the logger might be configured in various ways (e.g.
print to file). We want that the message ends up in the same file.

The default is also to use ktoolbox.common.logger. Callers won't have to
do anything, but they can provide their own logger (or set it to None).