common: change format for format_duration() to show days
Previously, if the duration was exact seconds, we would not show the
subseconds. This changes now, we always show 4 fractional digits. The
benefit is that the format is similarly styled regardless of the
duration.

Also, if there are large number of hours, it becomes hard for the human
to quickly calculate how many days this is (at least for me). It seems
preferable to show the number of days in that case. Obviously, if you
have 24 hours or more, then the format no longer strictly follows the
"HH:MM:SS.ssss" pattern.

You can argue whether this or another format is preferable. For now, I
think this is the "best", hence the change.