common: improve FutureThread.result() to better behave on cancel
When we cancel (or when we cancel_on_timeout), I think we should
not return from the function before the result was collected.

That is, if you call result() with cancel=True, then we should block
until we have the result. A subsequent result() should no longer block.

Also, add a FutureThread.result_full() variant. That is so that the
caller can distinguish between timeout and completion.