common: add "own_log_stram" argument to Serial() to close log_stream during close
We want to wrap the Serial() in a with statement. But we may also want
to have a create_serial() function that creates the instance for us.

Add a mechianism for the Serial() to own the log-stream. Otherwise, the
create_serial() function would also need to return the log-stream
separately, and the caller would need to take care to close them.

This lets the closing of the stream delegate to Serial().