common: in run_main() only log message for an actual Exception
A sys.exit() raises a SystemExit. That is more like a regular return
from the function, unlike other exceptional cases.

We make the choice that we want to log a message before calling
cleanup(). And we choose to only log that message on exceptions.

Maybe it could be useful to just always log a message that cleanup is
now happening, but then we get noisy output when users pass "--help" to
argparse-based tools, which previously logged:

    Got exception 0. Run cleanup first

Instead, for regular exits (including SystemExit) don't log the message.