common: add ImmutableDataclass helper class
We want immutable classes, so we use a frozen dataclass.

However, it is useful to allow for a limited form of mutability.
The ImmutableDataclass provides three methods for its subclasses:
_field_set_once(), _field_get() and _field_check().

_field_set_once() is used to set a key value, but fails if the key is
already set. _field_get() can be used to retrieve the value, but only if
it was previously set. _field_check() returns a tuple (value, exists) to
check if a field is initialized without raising an exception.

Such fields are clearly mutable, that is, they don't exist from the
beginning of when the object was initialized. However, any user will
only ever see one particular value (or get an exception if they try to
get it before it was initialized or if they try to initialize it twice).
This form of limited mutability is useful, because the state changes in
a controlled manner. Any user will not see an attribute change, and it
can only be set at one time.

Unit test by Claude Code.