common: don't sleep after Serial.send()
In basically all cases will a send() be followed by an expect(), which
already has a timeout for waiting for the response.

This additional sleep by default is unnecessay and unexpected. Drop
it.

For convenience, the parameter is still here, if you really want to
sleep afterwards in one go.