common: improve logging for cleanup with run_main() and CleanupList
On the one hand, we want to log a message that we are about to
run the cleanup. On the other hand, we don't want to log
a message (and spam the output) if there is nothing to cleanup.

Previously, if cleanup parameter was just a plain callable,
we didn't know whether there is anything to cleanup.

Now, the preferred use is to combine run_main() with CleanupList.
In that case, we can check whether there is anything to cleanup
and avoid logging the messsge.

Obviously, the proper use of this is that the caller would explicitly
issue a CleanupList.cleanup() manually. The cleanup parameter in
run_main() is only a fallback in case of an unexpected exception.