common: extend ImmutableDataclass with a _field_set() method
The main point of ImmutableDataclass is to be mostly(!) immutable.  A
mostly immutable class is still useful. While we should always prefer
truly immutable instances (that gets initialized upon construction),
sometimes we need to modify state. ImmutableDataclass allows to do that
in a controlled manner.

That is, at dedicated places, we still want to be able to add/modify
fields.

Previously, we could set fields via:

  - _field_set_once(): would only allow to initilize a field
  - _field_get() with a on_missing callback

Let's extend this. There is now _field_set() which can reset the value
at will. However, you have to opt-in to "allow_missing" and "allow_set",
thereby you control whether initialization, resetting and unsetting is
allowed.

For convenience, there is also a _field_unset() wrapper, which allows to
unset a filed (but fails if it is not set).