common: use reentrant RLock in ImmutableDataclass
Various places in ImmutableDataclass call external callbacks
while holding the lock. This is of course prone to deadlock,
but holding the lock while invoking the callback is still
preferable over releasing the lock (and loose the guarantee
to do things a single time).

However, using the non-reentrant lock makes this especially
problematic, because it means we cannot access another field
without deadlocking ourselves.

Avoid that by using RLock.