Fix oc commands string for bash with metacharacters (e.g. |)
There are cases where the command (e.g. exec) needs to use pipes (e.g. |
grep "example"). Currently this will fail with shlex.split because it
cannot handle shell pipelines or other complex shell metacharacters and
will create it as a character instead of an operator to connect multiple
commands. Instead we should use "sh -c" such that the command is a
single string that is encapsulated.

Signed-off-by: William Zhao <wizhao@redhat.com>